# Reduce runtime of OptimizationExample