--- a/OpenSim/Tests/OptimizationExample_Arm26/testOptimizationExample.cpp
+++ b/OpenSim/Tests/OptimizationExample_Arm26/testOptimizationExample.cpp
@@ -41,41 +41,49 @@ using namespace std;
 //const static double refControls[ARM26_DESIGN_SPACE_DIM]
 //                    = {0.01, 0.01, 0.01, 0.99, 0.99, 0.948937};
 // This is the result when Millard2012Equilibrium muscles are used with rigid
-// tendons.
+// tendons, with an initial guess of {0.01, 0.01, 0.01, 0.99, 0.99, 0.99}.
 const static double refControls[ARM26_DESIGN_SPACE_DIM]
-                    = {0.0525133, 0.01, 0.01, 0.606306, 0.99, 0.136519};
+    = {0.060, 0.021, 0.022, 0.99, 0.71, 0.30};
 
- 
+
 int main()
 {
-    try {   
-        Storage result("Arm26_noActivation_states.sto"), standard("std_Arm26_noActivation_states.sto");
-        CHECK_STORAGE_AGAINST_STANDARD(result, standard, Array<double>(0.01, 16), __FILE__, __LINE__, "Arm26 no activation states failed");
+    try {
+        Storage result("Arm26_noActivation_states.sto"),
+                standard("std_Arm26_noActivation_states.sto");
+        CHECK_STORAGE_AGAINST_STANDARD(result, standard, Array<double>(0.01, 16),
+            __FILE__, __LINE__, "Arm26 no activation states failed");
         cout << "Arm26 no activation states passed\n";
 
-        // Check the optimization result acheived at least a velocity of 5.43 m/s, and that the control values are within either 20% 
-        // of the reference values or 0.05 in absolute value.
-        ifstream resFile; 
-        resFile.open("Arm26_optimization_result"); 
-        ASSERT(resFile.is_open(), __FILE__, __LINE__, "Can't open optimization result file" );
+        // Ensure the optimization result acheived a velocity of at least
+        // REF_MAX_VEL, and that the control values are within either 20% of the
+        // reference values or 0.05 in absolute value.
+        ifstream resFile;
+        resFile.open("Arm26_optimization_result");
+        ASSERT(resFile.is_open(), __FILE__, __LINE__,
+               "Can't open optimization result file" );
 
         SimTK::Array_<double> resVec;
         for ( ; ; ) {
             double tmp;
-            resFile >> tmp;   
-            if (!resFile.good()) 
-                break; 
-            resVec.push_back(tmp); 
+            resFile >> tmp;
+            if (!resFile.good())
+                break;
+            resVec.push_back(tmp);
         }
-    
-        ASSERT(resVec.size() == ARM26_DESIGN_SPACE_DIM+1, __FILE__, __LINE__, "Optimization result size mismatch" );
+
+        ASSERT(resVec.size() == ARM26_DESIGN_SPACE_DIM+1, __FILE__, __LINE__,
+               "Optimization result size mismatch" );
         
-        // We don't enforce the optimizer to find precisedly the same local minimum for now.    
-        /*for (int i = 0; i < ARM26_DESIGN_SPACE_DIM-1; i++) {
-            ASSERT(fabs(resVec[i] - refControls[i])/refControls[i] < 0.2 || fabs(resVec[i] - refControls[i]) < 0.05, __FILE__, __LINE__, "Control value does not match reference" );
-        }*/
-        ASSERT(resVec[ARM26_DESIGN_SPACE_DIM] > REF_MAX_VEL, __FILE__, __LINE__, "Optimized velocity smaller than reference" );
-        
+        // Ensure the optimizer found a local minimum we expect.
+        for (int i = 0; i < ARM26_DESIGN_SPACE_DIM-1; ++i) {
+            ASSERT(fabs(resVec[i] - refControls[i])/refControls[i] < 0.2 ||
+                   fabs(resVec[i] - refControls[i]) < 0.05, __FILE__, __LINE__,
+                   "Control value does not match reference" );
+        }
+        ASSERT(resVec[ARM26_DESIGN_SPACE_DIM] > REF_MAX_VEL, __FILE__, __LINE__,
+               "Optimized velocity smaller than reference" );
+
         cout << "Arm26 optimization results passed\n";
     }
     catch (const Exception& e) {