… OpenSim/Tests/OptimizationExample_Arm26/std_Arm26_noActivation_states.sto
… OpenSim/Tests/OptimizationExample_Arm26/std_Arm26_optimized_states.sto