Addresses #394