@chrisdembia 2015-04-23
testOptimizationExample fails when comparing the results of the optimization to those stored in [this file](https://github.com/opensim-org/opensim-core/blob/master/OpenSim/Tests/OptimizationExample_Arm26/std_Arm26_noActivation_states.sto). Perhaps the file should just be updated, if the new results look reasonable?