@aseth1 2015-04-23
>testOptimizationExample fails when comparing the results of the optimization to those stored in this file. Perhaps this file should just be updated, if the new results look reasonable?

@tkuchida did you update the standard on your machine for the test to pass?