@tkuchida 2015-04-23
>Perhaps this file should just be updated, if the new results look reasonable?

In the old standard motion, the elbow hyperextends to -28° at t=0.25s. With rigid tendons, the elbow remains relatively straight, hyperextending to only -4.6° at t=0.25s.