@tkuchida 2015-04-23
Runtime of testOptimizationExampleRuns:

compiler | master | this branch
:---: | :---: | :---:
MSVC | 411.87s | 123.47s
clang | 539.17s | 132.28s
gcc | 788.05s | 227.29s