@chrisdembia 2015-04-23
This is somewhat not related to this PR, but is the pennation angle functionality that was previously tested through this test now being tested elsewhere?

Also, for this maximal task, I would have thought the control inputs would all either be 0.01 or 0.99-ish. However, two of the control inputs are somewhere in the middle.