@chrisdembia 2015-04-23
Would it make sense to use an initial guess that is closer to the solution?