@tkuchida 2015-04-23
>is the pennation angle functionality that was previously tested through this test now being tested elsewhere?

There's a test ensuring that the pennation model is operating correctly in isolation (testMuscleFixedWidthPennationModel), but there's no test ensuring that the pennation model is operating correctly once hooked up to a Millard2012EquilibriumMuscle. I've added this to the list of tasks in #366.