@aseth1 2015-04-23
The test that optimal solution is within 20% of the reference controls remains commented out. We should uncomment to make sure we are getting similar solutions on all platforms. 
@chrisdembia that is a great and simple solution. It may also results in more uniform solutions on the different OS' and not to mention whittle down unnecessary computation time for the test case.