@tkuchida 2015-04-23
>that is a great and simple solution.

@aseth1 Are you referring to the suggestion to use an initial guess that is closer to the solution?