@tkuchida 2015-04-23
Runtime of testOptimizationExampleRuns:

compiler | master | previous push | now
:---: | :---: | :---: | :---:
MSVC | 411.87s | 123.47s | 76.57s
clang | 539.17s | 132.28s | 105.02s
gcc | 788.05s | 227.29s | 103.20s