@sherm1 2015-04-23
> Runtime of testOptimizationExampleRuns

Can you post that again? I blinked and missed it.