@tkuchida 2015-04-24
New runtime of testOptimizationExampleRuns:

compiler | master | this branch
:---: | :---: | :---:
MSVC | 411.87s | 63.24s
clang | 539.17s | 99.59s
gcc | 788.05s | 111.42s

The test now (1) runs faster, and (2) checks the results of a simulation using the _optimal_ controls rather than the _initial_ controls.