@aseth1 2015-04-24
Looks good to me including visualizing the results in OpenSim. Thanks @tkuchida .